refactor(router): Move router components to router module
