+++ packages/client/src/modules/router/components/AuthenticatedApp/index.ts
@@ -1,0 +1,1 @@
+export { AuthenticatedApp } from "./AuthenticatedApp";
