+++ packages/client/src/modules/router/components/UnauthenticatedApp/index.ts
@@ -1,0 +1,1 @@
+export { UnauthenticatedApp } from "./UnauthenticatedApp";
